VideoSettings: add HDMI selector for Herelink
This is not based on any of the custom build infrastructure.

However, in my opinion, this is where the setting should go, both
because it's the obvious/intuitive place (in the video stream settings),
and because it is somewhat close  to how it is in the stock version.

I don't think having it as a separate button in the toolbar is quite
right. It would take up too much space of the limited screen real
estate.

In the best case we could somehow enable this using the custom build
infrastructure. I'm open for suggestions but I can't quite wrap my head
around how it would be done.

Signed-off-by: Julian Oes <julian@oes.ch>